backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.1
pydantic==2.6.0
pandas==2.0.3
numpy==1.24.4
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.3.0
joblib==1.3.2
openai==1.12.0
websockets==12.0
langchain==0.1.13
langchain-openai==0.1.1
langchain-community==0.0.29
langchain-core==0.1.33
httpx==0.27.2

regex
